+++ crawler/src/Robots.cpp
@@ -1,6 +1,7 @@
 #include "Robots.h"
 
 #include "Clock.h"
+#include "CrawlerMetrics.h"
 #include "Util.h"
 #include "http/Request.h"
 #include "http/RequestExecutor.h"
@@ -390,6 +391,8 @@ void RobotRulesCache::ProcessPendingRequests() {
         queuedFetches_.pop();
     }
 
+    InFlightRobotsRequestsMetric.Set(executor_.InFlightRequests());
+
     if (executor_.InFlightRequests() == 0) {
         return;
     }
@@ -416,6 +419,12 @@ void RobotRulesCache::ProcessPendingRequests() {
 }
 
 void RobotRulesCache::HandleRobotsResponse(const http::CompleteResponse& r) {
+    RobotsResponseCodesMetric
+        .WithLabels({
+            {"status", std::to_string(r.header.status)}
+    })
+        .Inc();
+
     auto canonicalHost = CanonicalizeHost(r.req.Url());
     SPDLOG_TRACE("successful robots.txt request: {}", canonicalHost.host);